pyproject.toml
[project]
name = "jupyterm"
license = {file = "LICENSE"}
readme = "README.md"
description = "TUI Jupyter notebook viewer"
authors = [
    {name = "Sidney Mau", email = "sidneymau@gmail.com" }
]
dynamic = ["version"]
dependencies = [
    "nbformat",
    "pillow",
    "textual",
]

[project.scripts]
jupyterm = "jupyterm:main"
